[add] 1. datetime count api 优化 2. rename QueryXxxxAsync API --> XxxxAsync API

+++ MyDAL/Impls/CountImpl.cs
@@ -17,7 +17,7 @@ namespace Yunyong.DataExchange.Impls
         {
         }
 
-        public async Task<long> CountAsync()
+        public async Task<int> CountAsync()
         {
             DC.Action = ActionEnum.Select;
 
@@ -27,10 +27,10 @@ namespace Yunyong.DataExchange.Impls
             DC.DPH.SetParameter();
             DC.Method = UiMethodEnum.CountAsync;
             DC.SqlProvider.GetSQL();
-            return await DC.DS.ExecuteScalarAsync<long>();
+            return await DC.DS.ExecuteScalarAsync<int>();
         }
 
-        public async Task<long> CountAsync<F>(Expression<Func<M, F>> propertyFunc)
+        public async Task<int> CountAsync<F>(Expression<Func<M, F>> propertyFunc)
         {
             DC.Action = ActionEnum.Select;
             var keyDic = DC.EH.FuncMFExpression(propertyFunc)[0];
@@ -41,7 +41,7 @@ namespace Yunyong.DataExchange.Impls
             DC.DPH.SetParameter();
             DC.Method = UiMethodEnum.CountAsync;
             DC.SqlProvider.GetSQL();
-            return await DC.DS.ExecuteScalarAsync<long>();
+            return await DC.DS.ExecuteScalarAsync<int>();
         }
     }
 
@@ -53,7 +53,7 @@ namespace Yunyong.DataExchange.Impls
         {
         }
 
-        public async Task<long> CountAsync()
+        public async Task<int> CountAsync()
         {
             DC.Action = ActionEnum.Select;
 
@@ -63,10 +63,10 @@ namespace Yunyong.DataExchange.Impls
             DC.DPH.SetParameter();
             DC.Method = UiMethodEnum.JoinCountAsync;
             DC.SqlProvider.GetSQL();
-            return await DC.DS.ExecuteScalarAsync<long>();
+            return await DC.DS.ExecuteScalarAsync<int>();
         }
 
-        public async Task<long> CountAsync<F>(Expression<Func<F>> propertyFunc)
+        public async Task<int> CountAsync<F>(Expression<Func<F>> propertyFunc)
         {
             DC.Action = ActionEnum.Select;
             var dic = DC.EH.FuncTExpression(propertyFunc)[0];
@@ -76,7 +76,7 @@ namespace Yunyong.DataExchange.Impls
             DC.DPH.SetParameter();
             DC.Method = UiMethodEnum.JoinCountAsync;
             DC.SqlProvider.GetSQL();
-            return await DC.DS.ExecuteScalarAsync<long>();
+            return await DC.DS.ExecuteScalarAsync<int>();
         }
     }
 }
